% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_basp_bbs_data}
\alias{example_basp_bbs_data}
\title{BBS route-level survey data for Baird's Sparrow}
\format{
## `example_basp_bbs_data`
A data frame (tibble) with 5,490 rows and 5 columns:
\describe{
  \item{english}{Common name of species}
  \item{year}{year in which BBS survey was conducted}
  \item{route}{BBS survey route}
  \item{longitude}{BBS survey route start location WGS 84 decimal degrees}
  \item{latitude}{BBS survey route start location WGS 84 decimal degrees}
  ...
}
}
\source{
<https://doi.org/10.5066/P9GS9K64>
}
\usage{
example_basp_bbs_data
}
\description{
BBS counts for all BBS survey events on routes on which
Baird's sparrow has been observed at some point 1966-2022.
dataset includes
}
\keyword{datasets}
