% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid-range.R
\name{grid_range}
\alias{grid_range}
\title{Title}
\usage{
grid_range(
  species = NULL,
  seasonal_range = "breeding",
  range_map = NULL,
  crs_equal_area_custom = NULL,
  coverage_grid_custom = NULL,
  resolution = "27km"
)
}
\arguments{
\item{species}{character English common name of bird species to
support range data download using eBirdst package}

\item{seasonal_range}{character default = breeding, season of
monitoring program data and if possible range map}

\item{range_map}{optional sf package simple features polygon layer defining
the species' seasonal range}

\item{crs_equal_area_custom}{optional crs numerical identifier of equal area
coordinate system}

\item{coverage_grid_custom}{optional sf package simple features polygon
layer, equal area grid within which to calculate coverage}

\item{resolution}{character default = 27km (low resolution), argument passed
to `ebirdst::load_ranges()` function}
}
\value{
A list
}
\description{
Title
}
\examples{
# requires an ebirdst access key
# see ?ebirdst::set_ebirdst_access_key

ex_range <- grid_range(species = "Baird's Sparrow")

}
