% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlay-range-data.R
\name{overlay_range_data}
\alias{overlay_range_data}
\title{Identify grid with monitoring data}
\usage{
overlay_range_data(
  range,
  survey_sites,
  sites = "site",
  years = "year",
  proportion_area_included = 0,
  minimum_n_years = 1,
  x_coord = "longitude",
  y_coord = "latitude",
  crs_site_coordinates = 4326,
  add_survey_sites_to_range = TRUE,
  largest = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{range}{list - output object from `grid_range()` function.}

\item{survey_sites}{required - two options: 1 - dataframe or 2 - map
Dataframe input option requires a dataframe with at least three columns
representing the site-names, x-coordinate, and y-coordinates or monitoring
sites. Dataframe may also include column to indicate the year of a survey
event, if the goal is to estimate annual survey coverage. Dataframe option
is only suitable if survey site locations can be treated as points (e.g.,
survey sites are small areas or linear features in relation to the gridcells
used to calculate coverage - 120km across in the default grid)
Map option requires a LINE or POLYGON object simple features object from
package `sf` that represents the survey units as areas or transects. Object
must have an identified coordinate reference system (crs), but it can be
either a geographic or projected crs because the function will transform it
to match the crs of the maps in argument `range`}

\item{sites}{character column name in `survey_sites` that identifies the
unique site names}

\item{years}{optional character column name in `survey_sites` that identifies
years each site was surveyed}

\item{proportion_area_included}{optional numeric [0-1] minimum proportion of
a grid cell that overlaps a species range for the cell to be considered part
of the species' range (for example, if set to 0.25, then only grid cells
with at least 25 percent of their area inside the species' range are included in the
calculations of coverage and range size. Default = 0.0. May be relevant for
very patchy species' ranges, or at range boundaries.}

\item{minimum_n_years}{optional integer, minimum number of years with surveys
to consider a grid-cell to be covered in the calculation of overall coverage.
Default is 1, which means even a single year with data results in coverage.}

\item{x_coord}{character - name of column that contains decimal degrees
format x-coordinate if `survey_sites` is a dataframe.}

\item{y_coord}{character - name of column that contains decimal degrees
format y-coordinate if `survey_sites` is a dataframe.}

\item{crs_site_coordinates}{one of (i) character: a string accepted by GDAL,
(ii) integer, a valid EPSG value (numeric), or (iii) an object of class crs.
Ignore if `survey_sites` is not a dataframe. Default = 4326 (WGS 84). Accepts
any valid input type allowed by `sf::st_crs`.}

\item{add_survey_sites_to_range}{logical - whether grid cells with surveys
but outside of the species' range should be included as part of the species
range. Default = TRUE. Assumes that `survey_sites` represents locations where
the species has been observed during monitoring, and not just locations of
multi-species surveys. Setting to FALSE may generate non-finite estimates of
coverage if positive values of area covered within regions with no area in
species' range}

\item{largest}{logical passed to `sf::st_join` function. Default is FALSE}

\item{quiet}{logical FALSE by default, set to TRUE to suppress warning and
progress messages}
}
\value{
A list with the following five objects

1.    coverage_map - map of the cells in the species' range that intersect the survey data in each year. If a year column was included with the input data, then this mapped information is repeated for each year of the data. If no year information was included, then this object will be identical to the cumulative_coverage_map

2.    cumulative_coverage_map - map of cells in the species' range that intersect survey data in at least 1 year (if default value of minimum_n_years was used).

3.    total_gridded_range_area - land area of gridded range.

4.    annual_coverage_estimate - dataframe of area and proportion of range covered in each year

5.    cumulative_coverage_estimate = dataframe (1-row) of area and proportion of range covered.
}
\description{
Identify grid with monitoring data
}
\examples{
# requires an ebirdst access key
# see ?ebirdst::set_ebirdst_access_key

ex_range <- grid_range(species = "Baird's Sparrow")
#ex_coverage <- overlay_range_data(ex_range,
 #                                   survey_sites = example_basp_bbs_data)

}
