% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional-summary.R
\name{regional_summary}
\alias{regional_summary}
\title{Title}
\usage{
regional_summary(coverage = NULL, regions = NULL, region_name = "country")
}
\arguments{
\item{coverage}{list output of `overlay_range_data`}

\item{regions}{Map - a POLYGON simple features object from
package `sf` that represents the spatial regions within which
to summarise the coverage. Object must have an identified coordinate
reference system (crs), but it can be any type because the function
will transform it to match the crs of the maps in argument `coverage`}

\item{region_name}{character column in `regions` object that contains
the unique region name}
}
\value{
A list
}
\description{
Title
}
\examples{
#not run


}
