% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strata-with-data.R
\name{strata_w_data}
\alias{strata_w_data}
\title{Identify strata with monitoring data}
\usage{
strata_w_data(survey_sites, coverage_strata_custom = NULL)
}
\arguments{
\item{survey_sites}{simple feature point layer package `sf` with points,
lines, or polygons representing the survey locations.}

\item{coverage_strata_custom}{optional simple feature point layer package `sf`of
polygons representing the coverage strata. Must include a column named
strata_name that represents a unique identifier for each polygon.}
}
\value{
a `sf` simple features dataframe of the coverage strata with a
logical column indicating if that strata has monitoring data (i.e., if
monitoring data exist, column == TRUE and that stratum is "covered")
}
\description{
`strata_w_data()` overlays the location of monitoring sites with
coverage strata
}
\examples{
# Using the example model for Pacific Wrens...

# Generate the continental and stratum indices

}
\concept{indices and trends functions}
