% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid-range.R
\name{grid_range}
\alias{grid_range}
\title{Title}
\usage{
grid_range(
  species = NULL,
  seasonal_range = "breeding",
  range_map = NULL,
  crs_equal_area_custom = NULL,
  coverage_grid_custom = NULL,
  resolution = "27km",
  quiet = FALSE
)
}
\arguments{
\item{species}{character English common name of bird species to
support range data download using eBirdst package}

\item{seasonal_range}{character default = breeding, season of range map to
download. Acceptable values depend on the species' range estimates available
from ebirdst, including: "breeding", "nonbreeding", "postbreeding_migration",
and "prebreeding_migration", for migratory species and "resident" for
non-migratory species.}

\item{range_map}{Not yet implemented: eventually optional user-loaded range map
e.g., sf package polygon layer defining the species' seasonal range}

\item{crs_equal_area_custom}{optional crs numerical identifier of equal area
coordinate system}

\item{coverage_grid_custom}{optional sf package simple features polygon
layer, equal area grid within which to calculate coverage}

\item{resolution}{character default = "27km" (low resolution), argument passed
to `ebirdst::load_ranges()` function. Alternate high resolution = "9km".}

\item{quiet}{logical FALSE by default, set to TRUE to suppress warning and
progress messages}
}
\value{
A list with four components:

1.  coverage_grid - a map representing the species' eBird range map intersected with a regular grid. A simple feature `sf` polygon object of the regular grid applied to the range map.  If the defaults were used in the `grid_range()` function, then this will be a regular hexagonal grid that covers the Western Hemisphere with individual hexagons measuring approximately 120 km between parallel sides. If a custom regular grid was supplied to the function, e.g., `grid_range(..., coverage_grid_custom = my_regular_grid)`, then this will reflect the resolution and extent of that custom grid. The attributes will include at least these four columns: a unique name for each grid cell (grid_cell_name), the land area of the grid cell in squared kilometers (area_km2), the land area of the grid cell that is covered by the species' range (area_km2_inrange), the proportion of the grid cell's land area covered by the species' range (proportion_in_range).

2.  range_map - a map. A simple feature sf polygon object that represents the downloaded eBird range map.

3.  range_area - numeric. The total land area within the species' eBird range map.

4.  range_area_gridded - numeric. The total land area of grid cells that intersect with the species' eBird range map.
}
\description{
Title
}
\examples{
# requires an ebirdst access key
# see ?ebirdst::set_ebirdst_access_key

ex_range <- grid_range(species = "Baird's Sparrow")

ex_non_breeding_range <- grid_range(species = "American Robin",
"non_breeding")

}
