% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional-summary.R
\name{regional_summary}
\alias{regional_summary}
\title{Title}
\usage{
regional_summary(
  coverage = NULL,
  regions = NULL,
  region_name = "country",
  quiet = FALSE,
  use_intersection = TRUE
)
}
\arguments{
\item{coverage}{list output of `overlay_range_data`}

\item{regions}{Map - a POLYGON simple features object from
package `sf` that represents the spatial regions within which
to summarise the coverage. Object must have an identified coordinate
reference system (crs), but it can be any type because the function
will transform it to match the crs of the maps in argument `coverage`}

\item{region_name}{character column in `regions` object that contains
the unique region name}

\item{quiet}{logical FALSE by default, set to TRUE to suppress warning and
progress messages}

\item{use_intersection}{logical TRUE by default. uses an intersection of the
coverage grid with the summary regions, so that regions with some small
portion of a coerage grid cell are included in the calculations. If set to
FALSE, faster summaries but uses a spatial join that allocates partial coverage
grid cells to the region with greatest overlap.}
}
\value{
A list with the following five objects

1.   regional_coverage_map - map of the annual coverage grid that includes an indicator of which region the grid cell is allocated to (based on spatial join).

2.    regional_cumulative_coverage_map - map of the cumulative coverage grid that includes an indicator column for region.

3.    regional_map - a map of the polygons used to define the regions.

4.    regional_annual_coverage_estimate - dataframe of area and proportion of range covered in each year plus an additional column indicating which region each grid cell belongs to.

5.    regional_cumulative_coverage_estimate - dataframe of area and proportion of range covered by region.
}
\description{
Title
}
\examples{
#not run


}
